requirements.txt
fastapi==0.96.0
loguru
numpy==1.23.1
opencv-python
transformers==4.44.2
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
imagehash
Pillow
uvicorn
python-multipart